Fix printf format specifiers and use size_t for indices
- getUniqueID: int32_t -> size_t for index parameter
- printf: %ld/%d -> %x for size_t channel numbers
- driver_USB: use size_t for loop counter
- Add stddef.h include for size_t

+++ src/driver_USB.c
@@ -113,7 +113,7 @@ void usbSetup(void) {
 size_t board_get_unique_id(uint8_t id[], size_t max_len) {
   (void)max_len;
   uint32_t *pId = (uint32_t *)id;
-  for (int32_t i = 0; i < 4; i++) {
+  for (size_t i = 0; i < 4; i++) {
     *pId++ = getUniqueID(i);
   }
   return 16;
